rename js files to match convention
