+++ frontend/src/App.js
@@ -8,30 +8,30 @@ import 'bootstrap-icons/font/bootstrap-icons.css';
 import './static/custom.css';
 
 import Login from "./templates/index";
-import ParentDashboard from './templates/parent_dashboard';
+import ParentDashboard from './templates/parent-dashboard';
 
 import Students from "./templates/students";
-import StudentsDetail from "./templates/students_detail";
-import StudentsEdit from "./templates/students_edit";
-// import StudentsDelete from "./templates/students_delete";
+import StudentsDetail from "./templates/students-detail";
+import StudentsEdit from "./templates/students-edit";
+// import StudentsDelete from "./templates/students-delete";
 
 import Schools from "./templates/schools";
-import SchoolsDetail from "./templates/schools_detail";
-import SchoolsCreate from "./templates/schools_create";
-import SchoolsEdit from "./templates/schools_edit";
-// import SchoolsDelete from "./templates/schools_delete";
+import SchoolsDetail from "./templates/schools-detail";
+import SchoolsCreate from "./templates/schools-create";
+import SchoolsEdit from "./templates/schools-edit";
+// import SchoolsDelete from "./templates/schools-delete";
 
 import Users from "./templates/users";
-import UsersDetail from "./templates/users_detail";
-import UsersCreate from "./templates/users_create";
-import UsersEdit from "./templates/users_edit";
-// import UsersDelete from "./templates/users_delete";
+import UsersDetail from "./templates/users-detail";
+import UsersCreate from "./templates/users-create";
+import UsersEdit from "./templates/users-edit";
+// import UsersDelete from "./templates/users-delete";
 
 import BusRoutes from "./templates/routes";
-import BusRoutesDetail from "./templates/routes_detail";
-import BusRoutesPlanner from "./templates/routes_planner";
-import BusRoutesEdit from "./templates/routes_edit";
-// import BusRoutesDelete from "./templates/routes_delete";
+import BusRoutesDetail from "./templates/routes-detail";
+import BusRoutesPlanner from "./templates/routes-planner";
+import BusRoutesEdit from "./templates/routes-edit";
+// import BusRoutesDelete from "./templates/routes-delete";
 
 class App extends Component {
   render() {
@@ -42,23 +42,23 @@ class App extends Component {
           <Route path="/dashboard" element={<ParentDashboard />} />
 
           <Route path="/students" element={<Students />} />
-          <Route path="/students_detail" element={<StudentsDetail />} />
-          <Route path="/students_edit" element={<StudentsEdit />} />
+          <Route path="/students-detail" element={<StudentsDetail />} />
+          <Route path="/students-edit" element={<StudentsEdit />} />
 
           <Route path="/schools" element={<Schools />} />
           <Route path="/schools/create" element={<SchoolsCreate />} />
-          <Route path="/schools_detail" element={<SchoolsDetail />} />
-          <Route path="/schools_edit" element={<SchoolsEdit />} />
+          <Route path="/schools-detail" element={<SchoolsDetail />} />
+          <Route path="/schools-edit" element={<SchoolsEdit />} />
 
           <Route path="/users" element={<Users />} />
           <Route path="/users/create" element={<UsersCreate />} />
-          <Route path="/users_detail" element={<UsersDetail />} />
-          <Route path="/users_edit" element={<UsersEdit />} />
+          <Route path="/users-detail" element={<UsersDetail />} />
+          <Route path="/users-edit" element={<UsersEdit />} />
 
           <Route path="/routes" element={<BusRoutes />} />
-          <Route path="/routes_detail" element={<BusRoutesDetail />} />
-          <Route path="/schools/routes_planner" element={<BusRoutesPlanner />} />
-          <Route path="/routes_edit" element={<BusRoutesEdit />} />
+          <Route path="/routes-detail" element={<BusRoutesDetail />} />
+          <Route path="/schools/routes-planner" element={<BusRoutesPlanner />} />
+          <Route path="/routes-edit" element={<BusRoutesEdit />} />
         </Routes>
       </div>
     );
