+++ frontend/src/templates/schools-create.js
