+++ frontend/src/templates/schools-detail.js
