+++ frontend/src/templates/students-detail.js
